update authRepImpl (getChar and saveChar) logic

+++ quest-maker/app/src/main/java/com/example/quest_maker/domain/repository/AuthorRepositoryInterface.java
@@ -7,6 +7,6 @@ public interface AuthorRepositoryInterface {
 
     public boolean saveCharacteristic(SaveCharacteristicParam param);
 
-    public byte getCharacteristic(Characteristic param);
+    public short getCharacteristic(Characteristic param);
 
 }
